debugged url db problem, changes imageService back to app. image service ready for mounting
