+++ webpack.config.js
@@ -4,7 +4,7 @@ const DIST_DIR = path.join(__dirname, 'client/dist');
 
 module.exports = {
   mode: 'development',
-  entry: `${SRC_DIR}/imageService.jsx`,
+  entry: `${SRC_DIR}/app.jsx`,
   output: {
     filename: 'bundle.js',
     path: DIST_DIR
